Track whether a room lifecycle operation is in progress
Remove the temporary hasOperationInProgress stored property added in
b6c3ddb.

It seems to me, at least based on what I’ve implemented so far, that you
have an operation in progress if and only if you’re in a certain status
case, so I’ve encoded this constraint in the Status type. Might turn out
this is false and needs to be revisited later.

TODO note it’s your responsibility to put yourself back into a
non-in-progress state (i.e. same as it would be to unlock a mutex)